% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNpde-plotFunctions.R
\name{npde.plot.data}
\alias{npde.plot.data}
\title{Plot a NpdeData object}
\usage{
npde.plot.data(npdeObject, ...)
}
\arguments{
\item{npdeObject}{an object returned by a call to \code{\link{npde}} or \code{\link{autonpde}}}

\item{\dots}{additional arguments to be passed on to the function, to control which metric (npde, pd, npd) is used or to override graphical parameters (see the PDF document for details, as well as \code{\link{set.plotoptions}})}
}
\description{
Produces a spaghetti plot of the data
}
\keyword{plot}
