% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNpde-plotFunctions.R
\name{npde.plot.npde}
\alias{npde.plot.npde}
\alias{npde.plot.pd}
\title{Plots for pd and npde}
\usage{
npde.plot.pd(npdeObject, ...)

npde.plot.npde(npdeObject, ...)
}
\arguments{
\item{npdeObject}{an object returned by a call to \code{\link{npde}} or \code{\link{autonpde}}}

\item{\dots}{additional arguments to be passed on to the function, to control which metric (npde, pd, npd) is used or to override graphical parameters (see the PDF document for details, as well as \code{\link{set.plotoptions}})}
}
\description{
Plots for pd and npde
}
\keyword{plot}
