% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeData-methods.R, R/NpdeRes-methods.R,
%   R/NpdeObject-methods.R
\name{print.NpdeData}
\alias{print.NpdeData}
\alias{print.NpdeRes}
\alias{print.NpdeObject}
\title{Prints objects from the npde package}
\usage{
\method{print}{NpdeData}(x, nlines = 10, ...)

\method{print}{NpdeRes}(x, nlines = 10, ...)

\method{print}{NpdeObject}(x, nlines = 10, ...)
}
\arguments{
\item{x}{an object of class NpdeData, NpdeSimData, NpdeRes or NpdeObject}

\item{nlines}{number of lines from the dataset to print}

\item{...}{Additional arguments (ignored)}
}
\description{
prints  objects of classes NpdeData, NpdeSimData, NpdeRes and NpdeObject
}
