% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R, R/NpdeData-methods.R
\docType{methods}
\name{read}
\alias{read}
\alias{read-methods}
\alias{read,NpdeData}
\alias{read,NpdeData-methods}
\alias{read,NpdeSimData-methods}
\alias{read,NpdeData-method}
\alias{read,NpdeSimData-method}
\title{Read data into structures (internal)}
\usage{
read(object, name.data, header=TRUE, sep="", na.strings=c("NA","."), detect=TRUE,
verbose=FALSE)

\S4method{read}{NpdeData}(
  object,
  name.data,
  header = TRUE,
  sep = "",
  na.strings = c("NA", "."),
  detect = TRUE,
  verbose = FALSE
)

\S4method{read}{NpdeSimData}(
  object,
  name.data,
  header = TRUE,
  sep = "",
  na.strings = c("NA", "."),
  verbose = FALSE
)
}
\arguments{
\item{object}{an object}

\item{name.data}{character string giving the name of the dataset (can be a file on disk or an R dataframe)}

\item{header}{boolean indicating whether the file has a header (mandatory if
detect is TRUE)}

\item{sep}{field separator (for files on disk)}

\item{na.strings}{strings to be considered as indicating NA}

\item{detect}{a boolean; if TRUE, automatic recognition of names will be attempted to detect necessary items (longitudinal data structure, missing data and censoring information)}

\item{verbose}{a boolean; messages are printed if verbose is TRUE (defaults to FALSE)}
}
\value{
an object of class \code{"\linkS4class{NpdeData}"} or \code{"\linkS4class{NpdeSimData}"}
}
\description{
Creates an NpdeData object containing the observed data, either from disk or from a dataframe
}
\keyword{internal}
\keyword{methods}
