% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdeControl.R
\name{replace.plotoptions}
\alias{replace.plotoptions}
\title{Replace graphical options}
\usage{
replace.plotoptions(plot.opt,...)
}
\arguments{
\item{plot.opt}{a list of graphical preferences}

\item{\dots}{names and values of the options to be replaced}
}
\value{
an updated list of options for graphs
}
\description{
This function is used to replace graph options (available in the prefs slot of the NpdeObject object) for plotting NpdeObject objects
}
\details{
See documentation for a list of available options. During replacement, invalid (including misspelled) options will raise warnings.
}
\seealso{
\code{\link{npde}}, \code{\link{autonpde}}
}
\author{
Emmanuelle Comets <emmanuelle.comets@bichat.inserm.fr>
}
\keyword{internal}
\keyword{plot}
