% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeData-methods.R, R/NpdeRes-methods.R,
%   R/NpdeObject-methods.R
\name{showall}
\alias{showall}
\alias{showall.NpdeData}
\alias{showall,NpdeData-method}
\alias{showall.default}
\alias{showall,method}
\alias{showall.NpdeRes}
\alias{showall.NpdeObject}
\title{Contents of an object}
\usage{
showall(object)

\method{showall}{NpdeRes}(object)

\method{showall}{NpdeObject}(object)
}
\arguments{
\item{object}{a NpdeData object}
}
\description{
Prints the contents of an object
}
\keyword{print}
