% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeData-methods.R, R/NpdeObject-methods.R
\name{subset.NpdeData}
\alias{subset.NpdeData}
\alias{subset.NpdeObject}
\title{Subsetting a NpdeData object}
\usage{
\method{subset}{NpdeData}(x, subset, ...)

\method{subset}{NpdeObject}(x, subset, ...)
}
\arguments{
\item{x}{A NpdeData object}

\item{subset}{logical expression indicating elements or rows to keep: missing values are taken as false.}

\item{...}{Additional arguments (ignored)}
}
\description{
Return subset of data from a NpdeData object
}
