% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeData-methods.R, R/NpdeRes-methods.R,
%   R/NpdeObject-methods.R
\name{summary.NpdeData}
\alias{summary.NpdeData}
\alias{summary.NpdeRes}
\alias{summary.NpdeObject}
\title{Summary of a NpdeData object}
\usage{
\method{summary}{NpdeData}(object, print = TRUE, ...)

\method{summary}{NpdeRes}(object, print = TRUE, ...)

\method{summary}{NpdeObject}(object, ...)
}
\arguments{
\item{object}{A NpdeData object}

\item{print}{whether to print to data to stdev}

\item{...}{Additional arguments (ignored)}
}
\description{
Extracts elements from a NpdeData object
}
